# OpenAPI接口.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "默认模块",
    "description": "",
    "version": "1.0.0"
  },
...
  "paths": {
...
    "/auth/login": {
      "post": {
        "summary": "登录",
        "deprecated": false,
        "description": "",
        "tags": [
          "账户相关"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  }
                },
                "required": [
                  "email",
                  "password"
                ]
              },
              "examples": {}
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer"
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "email": {
                          "type": "string"
                        },
                        "nickname": {
                          "type": "string"
                        },
                        "role": {
                          "type": "string"
                        },
                        "token": {
                          "type": "string"
                        },
                        "quotaSpace": {
                          "type": "integer"
                        },
                        "quotaCount": {
                          "type": "integer"
                        },
                        "usedSpace": {
                          "type": "integer"
                        },
                        "usedCount": {
                          "type": "integer"
                        }
                      },
                      "required": [
                        "email",
                        "nickname",
                        "role",
                        "token",
                        "quotaSpace",
                        "quotaCount",
                        "usedSpace",
                        "usedCount"
                      ]
                    }
                  },
                  "required": [
                    "code",
                    "message",
                    "data"
                  ]
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/auth/email/verification": {
      "get": {
        "summary": "获取邮箱验证码",
        "deprecated": false,
        "description": "",
        "tags": [
          "账户相关"
        ],
        "parameters": [
          {
            "name": "email",
            "in": "query",
            "description": "",
            "required": true,
            "example": "xingyuan_guo@outlook.com",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "type",
            "in": "query",
            "description": "",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer"
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "type": "null"
                    }
                  },
                  "required": [
                    "code",
                    "message",
                    "data"
                  ]
                },
                "example": {
                  "code": 1,
                  "message": "success",
                  "data": null
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/auth/forgetPassword": {
      "post": {
        "summary": "忘记密码",
        "deprecated": false,
        "description": "",
        "tags": [
          "账户相关"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  },
                  "verificationCode": {
                    "type": "string"
                  }
                },
                "required": [
                  "email",
                  "password",
                  "verificationCode"
                ]
              },
              "examples": {}
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer"
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "type": "null"
                    }
                  },
                  "required": [
                    "code",
                    "message",
                    "data"
                  ]
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/auth/register": {
      "post": {
        "summary": "注册",
        "deprecated": false,
        "description": "注册接口，需要输入邮箱验证码，后端会进行校验。校验成功后，会返回一个token，前端可以根据这个token登录。",
        "tags": [
          "账户相关"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "nickname": {
                    "type": "string"
                  },
                  "email": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  },
                  "verificationCode": {
                    "type": "string"
                  }
                },
                "required": [
                  "nickname",
                  "email",
                  "password",
                  "verificationCode"
                ]
              },
              "examples": {}
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer"
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "email": {
                          "type": "string"
                        },
                        "nickname": {
                          "type": "string"
                        },
                        "role": {
                          "type": "string"
                        },
                        "token": {
                          "type": "string"
                        },
                        "quotaSpace": {
                          "type": "integer"
                        },
                        "quotaCount": {
                          "type": "integer"
                        },
                        "usedSpace": {
                          "type": "integer"
                        },
                        "usedCount": {
                          "type": "integer"
                        }
                      },
                      "required": [
                        "email",
                        "nickname",
                        "role",
                        "token",
                        "quotaSpace",
                        "quotaCount",
                        "usedSpace",
                        "usedCount"
                      ]
                    }
                  },
                  "required": [
                    "code",
                    "message",
                    "data"
                  ]
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
...
  },
  "components": {
...
    "securitySchemes": {}
  },
  "servers": [],
  "security": []
}